[Upgrades] Move most of the inplace upgrade code to its own module
While doing this, make the --inplace-upgrade-* commands exit the
server instead of requiring `--bootstrap-only` to be specified.
